requirements.txt
Flask==3.0.3
scikit-learn==1.4.2
pandas==2.2.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.1.0
joblib==1.4.0
